requirements.txt
Django==4.2.3
gunicorn==21.2.0
gevent==23.9.1
psycopg2==2.9.9
django-ckeditor